requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.10

# Vector database
qdrant-client>=1.7.0

# Search
tavily-python>=0.3.0

# LLM integrations
langchain-ollama>=0.1.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
